requirements.txt
# Gap Analyzer - Job Description & Résumé Analyzer
groq>=0.4.0
pypdf>=4.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
flask>=3.0.0
jsonschema>=4.0.0
pytest>=7.0.0
